components: Set up initial components for ChatBar, MessageList (which holds Messages), NavBar.
ChatBar.jsx: Current just returns footer with class chatbar and two input fields.

MessageList.jsx: Renders an array of Message.jsx components and displays them in the main HTML helement with class messages. We pass in a username, content and key prop to the Message component.

Message.jsx: Currently returns a div that holds a div with class message, which contains message-username and message-content spans which are passed as props to this component

App.jsx: returns the NavBar, MessageList and ChatBar components passing the appropriate props to them

index.jsx: added a require to load the home.scss style